+++ test/utils/types.ts
@@ -1,12 +1,21 @@
 import { Database } from '@/database/database'
-import { State } from '@/data/types'
+import { Elements, ModulePath, State } from '@/data/types'
 import { Repository } from '@/repository/repository'
+
+export type Entities = {
+  [name: string]: Elements
+}
 
 export interface TestingStore {
   $database: Database
   $databases: Record<string, Database>
-  state: Record<string, State>
   $repo(modelOrRepository: any, connection?: string): Repository<any>
+
+  // for tests
+  getRootState(connection?: string): Record<string, any>
+  writeModule(path: ModulePath, data: Entities): void
+  hasModule(path: ModulePath): boolean
+  getModuleData(path: ModulePath): State
 }
 
 export type TestingStoreFactory = () => TestingStore