+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## 2023/11/24
+### Core:
+* add `getEntityNames` method to database instance
+
+### Vuex:
+* prevent VuexDataProvider from registering modules twice
+
 ## 2023/11/23
 * Project started